+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-## Upcoming
+## 1.2.0
+* Stamina and regen is now reduced by armor (2.5% per armor point)
 * Slowdown due to low / locked stamina is now applied to swimming too
 
 ## 1.1.3